requirements.txt
requests==2.*
discord.py==2.1.*
discord==1.*
PyYAML==6.0
